intl/l10n/test/dom/test_domloc_translateElements.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test DOMLocalization.prototype.translateElements</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  const { DOMLocalization } =
    ChromeUtils.import("resource://gre/modules/DOMLocalization.jsm", {});
  const { MessageContext } =
    ChromeUtils.import("resource://gre/modules/MessageContext.jsm", {});

  async function* mockGenerateMessages(locales, resourceIds) {
    const mc = new MessageContext(locales);
    mc.addMessages("title = Hello World");
    mc.addMessages("link\n    .title = Click me");
    yield mc;
  }

  window.onload = async function() {
    SimpleTest.waitForExplicitFinish();

    const domLoc = new DOMLocalization(
      [],
      mockGenerateMessages
    );

    const p1 = document.querySelectorAll("p")[0];
    const link1 = document.querySelectorAll("a")[0];

    await domLoc.translateElements([p1, link1]);

    is(p1.textContent, "Hello World");
    is(link1.getAttribute("title"), "Click me");

    SimpleTest.finish();
  };
  </script>
</head>
<body>
  <p data-l10n-id="title" />
  <a data-l10n-id="link" />
</body>
</html>
